requirements.txt
# Core AI & Orchestration
langgraph>=0.5.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
crewai>=0.80.0
ollama
# Web Search & Scraping
tavily>=0.3.0
duckduckgo-search>=7.3.0
trafilatura>=1.12.0
beautifulsoup4>=4.12.0
lxml>=5.3.0
requests>=2.32.0
httpx>=0.28.0
ddgs

# UI
gradio>=5.0.0
gradio-client>=1.0.0
streamlit

# Utilities
python-dotenv>=1.0.0
pydantic>=2.10.0
tiktoken>=0.8.0
aiohttp>=3.11.0
tenacity>=9.0.0

# Deployment
huggingface-hub>=0.28.0